Use the same length for all scene graph numeric data
Having one int between all the qint64 is really not worth it. Also, the
clients were not picking that up correctly.

+++ src/quick/util/qquickprofiler.cpp
@@ -88,7 +88,7 @@ void QQuickProfilerData::toByteArrays(QList<QByteArray> &messages) const
                     // RendererFrame: preprocessTime, updateTime, bindingTime, renderTime
                     case QQuickProfiler::SceneGraphRendererFrame: ds << subtime_1 << subtime_2 << subtime_3 << subtime_4; break;
                     // AdaptationLayerFrame: glyphCount (which is an integer), glyphRenderTime, glyphStoreTime
-                    case QQuickProfiler::SceneGraphAdaptationLayerFrame: ds << (int)subtime_1 << subtime_2 << subtime_3; break;
+                    case QQuickProfiler::SceneGraphAdaptationLayerFrame: ds << subtime_1 << subtime_2 << subtime_3; break;
                     // ContextFrame: compiling material time
                     case QQuickProfiler::SceneGraphContextFrame: ds << subtime_1; break;
                     // RenderLoop: syncTime, renderTime, swapTime
